WIP - Removed RegisterUser.js, moved API request function to Register.js; API request needs fixing (CORS issue)
