+++ frontend/src/admin/guestsPage/pages/Guests.js
@@ -3,10 +3,11 @@ import classes from "../components/Guests.module.css";
 import Test from "../server/test";
 
 function Guests() {
+  const data = Test();
   const cards = [
     {
-      // cardName: data.id,
-      // description: data.username,
+      cardName: data.id,
+      description: data.username,
     },
     {
       cardName: "Two",
